backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers
torch
torchvision
torchaudio
google-generativeai 